refactor: Setting/Refdata Endpoint
Changed location of refdata and settings endpoints
