+++ src/constants/config.js
@@ -1,7 +1,3 @@
-const REFDATA_ENDPOINT = 'oa/refdata';
-const SETTINGS_ENDPOINT = 'oa/settings/appSettings';
-export { REFDATA_ENDPOINT, SETTINGS_ENDPOINT };
-
 const PANE_DEFAULT_WIDTH = '50%';
 export { PANE_DEFAULT_WIDTH };
