--- src/constants/index.js
@@ -1,5 +1,0 @@
-export { default as publicationRequestFields } from './publicationRequestFields';
-export {
-  REFDATA_ENDPOINT,
-  SETTINGS_ENDPOINT
-} from './config';
